cleaned up subsystems
