--- src/main/java/frc/robot/subsystems/CargoIntakeSubsystem.java
@@ -1,41 +1,0 @@
-/*----------------------------------------------------------------------------*/
-/* Copyright (c) 2018 FIRST. All Rights Reserved.                             */
-/* Open Source Software - may be modified and shared by FRC teams. The code   */
-/* must be accompanied by the FIRST BSD license file in the root directory of */
-/* the project.                                                               */
-/*----------------------------------------------------------------------------*/
-
-package frc.robot.subsystems;
-
-import com.ctre.phoenix.motorcontrol.ControlMode;
-import com.ctre.phoenix.motorcontrol.can.TalonSRX;
-
-import edu.wpi.first.wpilibj.command.Subsystem;
-import frc.lib.util.Utils;
-import frc.robot.Robot;
-import frc.robot.commands.cargosubsystem.MoveBallCommand;
-
-/**
- * Add your docs here.
- */
-public class CargoIntakeSubsystem extends Subsystem {
-  // Put methods for controlling this subsystem
-  // here. Call these from Commands.
-  private TalonSRX intake_roller;
-
-  public CargoIntakeSubsystem() {
-    intake_roller = new TalonSRX(Robot.m_robotMapProperties.getIntakeRoller());
-  }
-
-  public void moveBall(double power) {
-    power = Utils.Deadband(power, 0.1);
-    intake_roller.set(ControlMode.PercentOutput, power);
-  }
-
-  @Override
-  public void initDefaultCommand() {
-    setDefaultCommand(new MoveBallCommand());
-    // Set the default command for a subsystem here.
-    // setDefaultCommand(new MySpecialCommand());
-  }
-}
